backend/utils/pdf_parser.py
import fitz  # PyMuPDF


def extract_text_from_pdf(file_bytes: bytes) -> str:
    """
    Extract clean text from PDF bytes using PyMuPDF.
    Handles multi-page PDFs and cleans whitespace.
    """
    try:
        doc = fitz.open(stream=file_bytes, filetype="pdf")
        text = ""
        for page in doc:
            text += page.get_text("text") + "\n"
        doc.close()

        lines = [line.strip() for line in text.splitlines() if line.strip()]
        return "\n".join(lines)
    except Exception as e:
        raise ValueError(f"PDF extraction failed: {str(e)}")
