backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
langchain==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.1.4
langchain-community==0.3.0
langchain-core==0.3.0
chromadb==0.5.0
openai==1.50.0
PyMuPDF==1.24.0
motor==3.5.0
pymongo==4.8.0
pydantic==2.9.0
python-dotenv==1.0.1
